+++ NEWS.md
@@ -1,3 +1,8 @@
+# nflplotR 1.6.0
+
+* Added new experimental function `gt_pct_bar()`. A helper that adds filled percentage bars to 'gt' table to add context to values. (#76, #77)
+* Changed css styling of images in package documentation by CRAN request. (#90)
+
 # nflplotR 1.5.0
 
 * Rewrite theme elements in S7 to fully work with ggplot2 v4. This means nflplotR now requires ggplot2 v4! nflplotR is a ggplot2 extension and if ggplot2 version jumps make breaking changes, then it's best for the extension to not try to be downwards compatible. The theme system is now fully powered by ggpath. (#73)
@@ -5,7 +10,7 @@
 * The theme elements `element_nfl_logo()` and `element_nfl_wordmark()` now clean team abbreviations by calling `nflreadr::clean_team_abbrs()` and insert empty grobs for mismatches. 
 * All geoms and theme elements will print more informative warnings about team abbreviation, or player ID mismatches.
 * The gt logo rendering functions now add alt text for accessibility and to silence an annoying pkgdown warning. (#69)
-* nflplotR now requires R 4.1 because magick needs this R version. This also follows the [Tidyverse R version support rules](https://www.tidyverse.org/blog/2019/04/r-version-support/). (#71)
+* nflplotR now requires R 4.1 because magick needs this R version. This also follows the [Tidyverse R version support rules](https://tidyverse.org/blog/2019/04/r-version-support/). (#71)
 
 # nflplotR 1.4.0